requirements.txt
jsonfield

# Additional requirements go here
social-auth-app-django
asgiref==3.8.0
O365
azure-core==1.30.1
azure-storage-blob==12.19.1
dateparser==1.2.0